requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Quality metrics
lpips>=0.1.4
pytorch-msssim>=1.0.0
torchmetrics>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
